% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{check_for_existence}
\alias{check_for_existence}
\title{check_for_existence}
\usage{
check_for_existence(check_here, the_dataset, the_column)
}
\arguments{
\item{check_here}{A folder on your computer to save the PDFs to, for instance "/Desktop/arxiv"}

\item{the_dataset}{The tibble that has the URLs that you'll be getting and the full filenames
that you'd like to save them to.}

\item{the_column}{The column that contains the full filenames that you'd like
to save them to. This needs to be a character vector.}
}
\description{
The \code{check_for_existence} function looks at the folder that you're
going to save your PDFs to and checks whether you've already got any of them. It
then suggests that you filter to remove them.
}
\examples{
two_pdfs <-
 tibble::tibble(
  locations_are = c("https://osf.io/preprints/socarxiv/z4qg9/download",
    "https://osf.io/preprints/socarxiv/a29h8/download"),
  save_here = c("~/competing_effects_on_the_average_age_of_infant_death.pdf",
     "~/cesr_an_r_package_for_the_canadian_election_study.pdf")
   )
get_these_and_save_them(two_pdfs$locations_are, two_pdfs$save_here)

check_for_existence("~", two_pdfs, "save_here")

}
