% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{get_these_and_save_them}
\alias{get_these_and_save_them}
\title{get_these_and_save_them}
\usage{
get_these_and_save_them(get_them_from_here, put_them_here)
}
\arguments{
\item{get_them_from_here}{A bunch of URLs where the PDFs are, usually a column.}

\item{put_them_here}{A location on your computer to save the PDF to.}
}
\description{
The \code{download_pdfs} function works with one PDF takes a tibble of locations (usually
URLs) and file names, and then downloads the PDF from the location to the file
name, saving as it goes. It politely waits 10 seconds between calls to the
location, skips locations that give an error and (optionally) prints a status
message. It will also (optionally) let you know about any PDFs that are very
small (which usually indicates there's something wrong with them).
}
\examples{
two_pdfs <-
 tibble::tibble(
  locations_are = c("https://osf.io/preprints/socarxiv/z4qg9/download",
    "https://osf.io/preprints/socarxiv/a29h8/download"),
  save_here = c("~/competing_effects_on_the_average_age_of_infant_death.pdf",
     "~/cesr_an_r_package_for_the_canadian_election_study.pdf")
   )
get_these_and_save_them(two_pdfs$locations_are, two_pdfs$save_here)

}
