% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{get_this_and_save_it}
\alias{get_this_and_save_it}
\title{get_this_and_save_it}
\usage{
get_this_and_save_it(get_it_from_here, put_it_here)
}
\arguments{
\item{get_it_from_here}{A URL of a PDF.}

\item{put_it_here}{A location on your computer to save the PDF to.}
}
\description{
The \code{get_this_and_save_it} function works with one PDF, taking the URL
of where that PDF is saved - get_it_from_here - and saving it to your computer at the
location that you specify in put_it_here. It prints a status to let you know
that it's done.
}
\examples{
from_here  <- c("https://osf.io/preprints/socarxiv/z4qg9/download")
put_it_here  <- c("~/competing_effects_on_the_average_age_of_infant_death.pdf")

get_this_and_save_it(from_here, put_it_here)

}
