% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{get_these_and_save_them}
\alias{get_these_and_save_them}
\title{get_these_and_save_them}
\usage{
get_these_and_save_them(
  get_them_from_here,
  put_them_here_as_this,
  wait_this_long = 7
)
}
\arguments{
\item{get_them_from_here}{A column of URLs from a tibble in character format.}

\item{put_them_here_as_this}{A column of filenames from a tibble that includes the full
path, in character format, that specify where and what you would like to save
the PDF as.}

\item{wait_this_long}{integer, that indicates the number of seconds to wait
between the requests. Automatically adds a bit of noise to lessen the effect
on systematic processes that might be otherwise working.}
}
\description{
The \code{get_these_and_save_them} function works with a tibble of
locations (usually URLs) and file names, and then downloads the PDF from the
location to the file name, saving as it goes, and letting you know where it's
up to. It politely waits around 7 seconds between calls to the
location, and skips locations that give an error.
}
\examples{
two_pdfs <-
 tibble::tibble(
  locations_are = c("https://osf.io/preprints/socarxiv/z4qg9/download",
    "https://osf.io/preprints/socarxiv/a29h8/download"),
  save_here = c("~/competing_effects_on_the_average_age_of_infant_death.pdf",
     "~/cesr_an_r_package_for_the_canadian_election_study.pdf")
   )
get_these_and_save_them(two_pdfs$locations_are, two_pdfs$save_here)

}
